Updated to the K4AdotNet package to allow multiple camera body tracking
Working version with just the Kinect Data capture and a simple Skeleton rendering for multiple cameras. Checked with 2 Femtos.
Need to update the Readme to refer to the new libraries.
Next step - (a) Perform Skeleton based Calibration of multiple cameras and apply that transformations from files to update all the Kinect data (b) Re-integrate the Point Cloud and Mesh part for both the entire scene + person.